+++ backend/SmartGarage/src/main/java/com/example/demo/repositories/PackRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PackRepository extends JpaRepository<Pack, Integer>{
+public interface PackRepository extends JpaRepository<Pack, Long>{
     boolean existsPacksByPackName(String packName);
 }
