+++ backend/SmartGarage/src/main/java/com/example/demo/service/RoleService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface RoleService {
     List<Role> getAllRoles();
     Role getRoleByRoleName(String roleName) throws Exception;
-    Role getRoleById(Integer roleId);
+    Role getRoleById(Long roleId);
 }
