+++ backend/SmartGarage/src/main/java/com/example/demo/service/UserService.java
@@ -14,16 +14,16 @@ public interface UserService extends UserDetailsService {
     RegistrationResponse register(User user, UserDTO request);
 
     List<User> getAllUsers(MultiValueMap<String, String> allParams);
-    Optional<User> getUserById(int userId);
+    Optional<User> getUserById(Long userId);
     Optional<User> getUserByUsername(String username);
     Optional<User> getUserByEmail(String email);
 
 //    List<User> searchByName(String name);
 
-    User updateUser(User user, int userId, UserDTO userDTO);
+    User updateUser(User user, Long userId, UserDTO userDTO);
 
     void changePassword(User user, String oldPassword, String newPassword);
 
-    void deleteUser(User user, int userId);
-    void deleteUsers(User user, List<Integer> ids);
+    void deleteUser(User user, Long userId);
+    void deleteUsers(User user, List<Long> ids);
 }
