+++ backend/SmartGarage/src/main/java/com/example/demo/service/VisitService.java
@@ -12,13 +12,13 @@ import java.util.Optional;
 public interface VisitService {
         List<Visit> getAllVisits(MultiValueMap<String, String> allParams);
 
-        Optional<Visit> getVisitById(int visitId);
-        List<Visit> getVisitsByVehicleId(int vehicleId);
+        Optional<Visit> getVisitById(Long visitId);
+        List<Visit> getVisitsByVehicleId(Long vehicleId);
         List<Visit> getVisitsByDate(LocalDateTime dateTime);
 
         Visit createVisit(User user, Visit visit);
-        Visit update(User user, int visitId, Visit changes);
+        Visit update(User user, Long visitId, Visit changes);
 
-        void deleteVisit(User user ,int visitId);
-        void deleteVisits(User user, List<Integer> ids);
+        void deleteVisit(User user ,Long visitId);
+        void deleteVisits(User user, List<Long> ids);
 }
